generalise handlers for task modals (other than submitting it) so these functions can be reused across add new task modal as well as edit existing task modal
